chore(linux): Add script to upload to debian
- Create and sign source package
- Upload to mentors.debian.net
- Create (and optionally push) commit with updated changelog file
  based on stable branch
- Create (and optionally push) commit with updated changelog file
  based on master branch

This change also slightly modifies `echo_heading` and outputs a
mark. This will show to the left of the line as well as in the
scroll bar in a terminal in VSCode [[1]] and will help to find the
headings if you have a lot of output in between.

[1]:https://code.visualstudio.com/updates/v1_69#_setmark-sequence-support

(cherry picked from commits ae61410cc93c47a785dbbd30d917043214a9f35b,
2a55ffb024dbc4d19e043dc93eb12273ed05d9f0, 9232c205694889345da27d690dce9dc05501326e,
420cc0814d57fe334c7e2a38efb5b7548ff078d2, 9f7b85834a6b50d4bd1ed26e79c85b44f6edd51d,
and ae0378a57fd2c7e53e5b1b7b63ecf435fd01d42d)