chore(mac): update setup/build to include build-utils.sh
shellHelperFunctions.sh currently needs build-utils.sh sourced first in
order to parse. As we are planning to merge these two scripts, I am okay
with this for now. But we need to make sure that prerequisite is met for
now.

I ran the following command to find scripts that referenced
shellHelperFunctions.sh but not build-utils.sh:

  grep -L build-utils.sh $(grep -lR shellHelperFunctions --include='*.sh' .)

The following files were found:

  ./mac/bashHelperFunctions.sh
  ./mac/setup/build.sh
  ./resources/build/history/history-utils.sh

* bashHelperFunctions.sh has only a comment reference, so it's okay.
* mac/setup/build.sh I have reworked to include build-utils.sh.
* history-utils.sh is used only by ios/tools/prepRelease.sh, and that
  script already includes build-utils.sh.