chore: Change platform advocates per discussion
As discussed, we are changing some of the platform advocates for (a)
geographical, and (b) platform familiarity reasons. This will be
reflected in the automatic assignment of reviewers on PRs.

We are pretty much all affected in this shuffle!

@keymanapp-test-bot skip

(cherry picked from commit 9bff83fa00df2be69e28e1833a1c29cebdcc90e3)