chore:  try disabling concurrency for browserstack tests
We're still getting sporadic test failures with BrowserStack. So per one
discussion online, attempting to run without concurrency to see if it
improves stability.

I note that the video from tests where we "fail to capture" shows a
browser with an error: "Safari Can't Connect to the Server" "Safari
can't open the page "localhost:9879/?id=75185158" because Safari can't
connect to the server "localhost".

https://automate.browserstack.com/dashboard/v2/builds/419c59decb8e12303e45e1f4bad1b945252df532/sessions/19712158833ce57592b66a50433965e4146fc972