chore(linux): Adjust Linux specific files in `.gitignore`
During Linux packaging some files get put in the root of the
source directory. This change limits these files in `.gitignore`
to the root directory. It would be better to not generate these files
in the root directory in the first place, but that's not trivial
and a separate task.