feat(windows): remove hotkey from pop-up. content-frame size
The hotkey option and editing did not need to be on the add remove
languages pop-up.
There size of the contentframe for all the tabs was incorrectly
calculated. The header size was not correct as the header had
padding. This caused the size to be redrawn incorrectly when
a scroll bar was first added.