fix(developer): Use US base layout for debugger
Fixes #7529.

We must always use the US base layout for the debugger, as positional
keyboards are not affected by the active Windows system keyboard. This
fix ensures that the behaviour in the debugger matches what the user
experiences in Keyman for Windows.

Note that this does not address support for mnemonic layouts, tracked in
issue #1074.