fix(developer): update monaco editor to 0.15.6
Updates Monaco editor from 0.14.3 to 0.15.6, incorporating
microsoft/vscode#57617 to fix #7625.

The changes to editor.js are required because `editor.model` was not a
public property and we should have been using `editor.getModel()` all
along.

I have not updated to the latest version of monaco (0.35.1 at time of
writing this), as I wish to minimize the changeset during beta.

While I prefer not to update dependencies in beta, this seems to be a
fairly clean update. Will monitor and can roll back if it leads to
instability.