fix(linux): set DEB_VERSION
DEB_VERSION is used to set `pkgversion` in the
`override_dh_auto_build` build step. When we moved to the
consolidated source package and thus could no longer use the
default debhelper build tools we lost the automatic setting of
DEB_VERSION, so this change adds the relevant definitions.

Fixes #7724.