fix(android): bump version_code major calculation
Version code is, for 17.1.33: 1810033. This supports up to 9999
builds for a given major.minor version. We only support up to 9 minor
versions for a given major version but at present we are not using
minor versions at all.

Despite this calculation, the version_code should be treated as an
opaque integer, used only for Play Store publishing.

NOTE: accidental publish of beta 17.0.6 during 16.0 beta cycle has
forced a version_major bump, which is a permanent change. Relates to
issue #7705.