fix(ios): use mobile mode for keyboard download pages
Fixes #7915.

On iPadOS 16, it seems that WKWebView defaults to lying about its user
agent to servers. We can switch this behaviour by using .mobile for the
preferredContentMode.

We should also update the website to work around this, so that users
always get the right downloads for their device, but the website
currently uses server-side determination of user agent for this
particular aspect of the keyboard search, so that's a much more
significant refactor, to make it client-side, with greater risk and much
more testing required.